+++ spec/lib/tracer_factory_spec.rb
@@ -3,8 +3,8 @@
 
 require 'rack/mock'
 require 'spec_helper'
-require 'zipkin-tracer/zipkin_json_tracer'
-require 'zipkin-tracer/zipkin_null_tracer'
+require 'zipkin-tracer/zipkin_http_sender'
+require 'zipkin-tracer/zipkin_null_sender'
 
 describe ZipkinTracer::TracerFactory do
   def middleware(app, config={})
@@ -18,24 +18,23 @@ describe ZipkinTracer::TracerFactory do
 
   let(:logger) { Logger.new(nil) }
   let(:subject) { described_class.new(config) }
-  let(:tracer) { double('Trace::NullTracer') }
+  let(:tracer) { double('Trace::NullSender') }
 
   before do
     allow(ZipkinTracer::Application).to receive(:logger).and_return(logger)
   end
 
   describe 'initializer' do
-    # see spec/lib/zipkin_kafka_tracer_spec.rb
+    # see spec/lib/zipkin_kafka_sender_spec.rb
     if RUBY_PLATFORM == 'java'
       context 'configured to use kafka', platform: :java do
-        require 'zipkin-tracer/zipkin_kafka_tracer'
+        require 'zipkin-tracer/zipkin_kafka_sender'
 
         let(:zookeeper) { 'localhost:2181' }
-        let(:zipkinKafkaTracer) { double('ZipkinKafkaTracer') }
         let(:config) { configuration(zookeeper: zookeeper) }
 
-        it 'creates a zipkin kafka tracer' do
-          allow(Trace::ZipkinKafkaTracer).to receive(:new) { tracer }
+        it 'creates a zipkin kafka sender' do
+          allow(Trace::ZipkinKafkaSender).to receive(:new) { tracer }
           expect(Trace).to receive(:tracer=).with(tracer)
           expect(described_class.new.tracer(config)).to eq(tracer)
         end
@@ -46,7 +45,7 @@ describe ZipkinTracer::TracerFactory do
       let(:config) { configuration(json_api_host: 'fake_json_api_host') }
 
       it 'creates a zipkin json tracer' do
-        allow(Trace::ZipkinJsonTracer).to receive(:new) { tracer }
+        allow(Trace::ZipkinHttpSender).to receive(:new) { tracer }
         expect(Trace).to receive(:tracer=).with(tracer)
         expect(described_class.new.tracer(config)).to eq(tracer)
       end
@@ -56,12 +55,21 @@ describe ZipkinTracer::TracerFactory do
       let(:config) { configuration(log_tracing: true) }
 
       it 'creates a logger tracer' do
-        allow(Trace::ZipkinLoggerTracer).to receive(:new) { tracer }
+        allow(Trace::ZipkinLoggerSender).to receive(:new) { tracer }
         expect(Trace).to receive(:tracer=).with(tracer)
         expect(described_class.new.tracer(config)).to eq(tracer)
       end
     end
 
+    context 'configured to use Amazon SQS' do
+      let(:config) { configuration(sqs_queue_name: 'zipkin-sqs') }
+
+      it 'creates a sqs tracer' do
+        allow(Trace::ZipkinSqsSender).to receive(:new) { tracer }
+        expect(Trace).to receive(:tracer=).with(tracer)
+        expect(described_class.new.tracer(config)).to eq(tracer)
+      end
+    end
 
     context 'no transport configured' do
       it 'creates a null tracer' do
@@ -70,10 +78,11 @@ describe ZipkinTracer::TracerFactory do
           { json_api_host: nil },
           { json_api_host: "" },
           { json_api_host: "\n\t 　\r" },
-          { zookeeper: "" }
+          { zookeeper: "" },
+          { sqs_queue_name: "" }
         ].each do |options|
           config = configuration(options)
-          allow(Trace::NullTracer).to receive(:new) { tracer }
+          allow(Trace::NullSender).to receive(:new) { tracer }
           expect(Trace).to receive(:tracer=).with(tracer)
           expect(described_class.new.tracer(config)).to eq(tracer)
         end